pyproject.toml
[project]
name = "google-scholar-alert"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "click==8.1.8",
    "crossref-commons==0.0.7",
    "dependency-injector==4.45.0",
    "extract-msg==0.52.0",
    "httpx==0.28.1",
    "pika==1.3.2",
    "pymongo==4.11",
    "pymupdf==1.25.3",
    "python-dotenv==1.0.1",
    "requests-toolbelt==1.0.0",
    "selenium==4.29.0",
]
